Refactor pt3
