+++ src/main/java/me/kaotich00/fwwar/war/assault/SiegeWar.java
@@ -102,11 +102,6 @@ public class SiegeWar extends AssaultWar {
     }
 
     @Override
-    public boolean supportKits() {
-        return false;
-    }
-
-    @Override
     public void handlePlayerDeath(Player player) {
     }
