+++ src/main/java/me/kaotich00/fwwar/war/bolt/RandomFactionWar.java
@@ -14,6 +14,7 @@ import me.kaotich00.fwwar.objects.kit.Kit;
 import me.kaotich00.fwwar.objects.war.ParticipantNation;
 import me.kaotich00.fwwar.objects.war.ParticipantTown;
 import me.kaotich00.fwwar.services.SimpleArenaService;
+import me.kaotich00.fwwar.services.SimpleKitService;
 import me.kaotich00.fwwar.services.SimpleScoreboardService;
 import me.kaotich00.fwwar.services.SimpleWarService;
 import me.kaotich00.fwwar.utils.*;
@@ -51,13 +52,11 @@ public class RandomFactionWar extends BoltWar {
     @Override
     public void startWar() {
         try {
-            Random random = new Random();
-            SimpleArenaService arenaService = SimpleArenaService.getInstance();
-            Arena warArena = arenaService.getArenas().get(arenaService.getArenas().size() > 1 ? random.nextInt(arenaService.getArenas().size() - 1) : 0);
+            Arena warArena = SimpleArenaService.getInstance().getRandomArena();
 
             setWarStatus(WarStatus.STARTED);
 
-            Kit kit = generateRandomKit();
+            Kit kit = SimpleKitService.getInstance().generateRandomKit();
 
             Nation firstNation = null;
             Map<UUID, Location> playersToTeleport = new HashMap<>();
@@ -154,32 +153,6 @@ public class RandomFactionWar extends BoltWar {
     }
 
     @Override
-    public void stopWar() {
-        for(ParticipantNation participantNation: this.getParticipants()) {
-            for(ParticipantTown participantTown: participantNation.getTowns()) {
-                Set<UUID> residents = participantTown.getPlayers();
-                Town town = participantTown.getTown();
-
-                for(UUID uuid: residents) {
-                    Player player = Bukkit.getPlayer(uuid);
-                    if(player != null) {
-                        player.getInventory().clear();
-                        player.getInventory().setArmorContents(null);
-
-                        try {
-                            player.teleport(town.getSpawn());
-                        } catch (TownyException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                }
-            }
-        }
-
-        setWarStatus(WarStatus.ENDED);
-    }
-
-    @Override
     public boolean supportKits() {
         return true;
     }
@@ -192,50 +165,6 @@ public class RandomFactionWar extends BoltWar {
     @Override
     public Optional<Kit> getPlayerKit(Player player) {
         return Optional.ofNullable(this.playerKits.get(player.getUniqueId()));
-    }
-
-    @Override
-    public void handlePlayerDeath(Player player) {
-        getDeathQueue().addPlayer(player);
-
-        try {
-            TownyAPI townyAPI = TownyAPI.getInstance();
-
-            Resident resident = townyAPI.getDataSource().getResident(player.getName());
-            Town town = resident.getTown();
-
-            Set<UUID> participants = null;
-            if(hasTown(town))
-                participants = getParticipant(town.getNation().getUuid()).getTown(town.getUuid()).getPlayers();
-
-            if(participants == null) {
-                return;
-            }
-
-            if (!participants.contains(player.getUniqueId())) {
-                return;
-            }
-
-            getParticipant(town.getNation().getUuid()).getTown(town.getUuid()).removePlayer(player.getUniqueId());
-            Message.WAR_PLAYER_DEFEATED.send(player);
-
-            boolean shouldRemoveNation = true;
-            if(getParticipant(town.getNation().getUuid()).getTowns().size() > 0){
-                shouldRemoveNation = false;
-            }
-
-            if(shouldRemoveNation) {
-                removeParticipant(town.getNation());
-                Message.NATION_DEFEATED.broadcast(town.getNation().getName());
-            }
-
-            if(getParticipants().size() < 2) {
-                SimpleScoreboardService.getInstance().removeScoreboards();
-                SimpleWarService.getInstance().stopWar();
-            }
-
-        } catch (TownyException ignored) {
-        }
     }
 
 }
